server.py
#!/usr/bin/env python3
"""
Yandex MCP Server

A Model Context Protocol (MCP) server providing AI assistants with access to
Yandex services through standardized tools and resources.

This server enables AI assistants to interact with Yandex APIs using OAuth 2.0
authentication. It implements the MCP specification for stdio-based communication.

License: GNU General Public License v3.0 (GPLv3)
"""

import os
import sys
from typing import Any

from dotenv import load_dotenv


def main() -> int:
    """
    Main entry point for the Yandex MCP server.
    
    This placeholder will be replaced with the full MCP server implementation
    in subsequent development phases.
    
    Returns:
        Exit code (0 for success, non-zero for errors)
    """
    load_dotenv()
    
    yandex_token = os.getenv("YANDEX_OAUTH_TOKEN")
    
    if not yandex_token:
        print(
            "Error: YANDEX_OAUTH_TOKEN environment variable is not set.",
            file=sys.stderr,
        )
        print(
            "Please set it in your .env file or environment.",
            file=sys.stderr,
        )
        return 1
    
    print("Yandex MCP Server (placeholder)", file=sys.stderr)
    print("Python 3.10+ • Licensed under GPLv3", file=sys.stderr)
    print("", file=sys.stderr)
    print("Server initialization will be implemented in future development.", file=sys.stderr)
    print("This is a placeholder for project bootstrapping.", file=sys.stderr)
    
    return 0


if __name__ == "__main__":
    sys.exit(main())
