requirements.txt
# Yandex MCP Server - Python Dependencies
# Python 3.10+ required

# MCP Framework
fastmcp>=2.0

# HTTP Client
httpx>=0.27

# Data Validation
pydantic>=2.0

# Environment Configuration
python-dotenv

# Testing Dependencies
pytest>=7.0
pytest-asyncio>=0.21
pytest-cov>=4.0
pytest-mock>=3.10
